Refactoring for the PR
