+++ src/NuGet.Clients/NuGet.CommandLine/Common/LocalResourceUtils.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Threading;
 
-namespace NuGet.CommandLine
+namespace NuGet.Common
 {
     public class LocalResourceUtils
     {
